video-merger/requirements.txt
# FFmpeg는 별도 설치 필요
# brew install ffmpeg (Mac) 또는 https://ffmpeg.org/download.html

# 자동 업로드 관련 패키지
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
playwright>=1.40.0
schedule>=1.2.0
python-dotenv>=1.0.0

# 웹 인터페이스
flask>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# AI 클리핑 시스템
yt-dlp>=2023.11.16
openai-whisper>=20231117
google-generativeai>=0.3.0
pydantic>=2.5.0
opencv-python>=4.8.0
numpy>=1.24.0
